using `darkTheme` and `defaultTheme` from `@dzcode.io/ui-mobile` package
