+++ mobile/src/screens/faq/styles.ts
@@ -1,5 +1,5 @@
+import { Colors } from "@dzcode.io/ui-mobile/dist/styles/colors";
 import { StyleSheet } from "react-native";
-import { Colors } from "src/styles/colors";
 import { MEDIUM_PADDING_SIZE, MEDIUM_TEXT_SIZE } from "src/utils/constants";
 
 export const faqStyles = StyleSheet.create({
