+++ mobile/src/screens/projects/card-item/index.tsx
@@ -1,8 +1,8 @@
 import { Project } from "@dzcode.io/api/dist/app/types/legacy";
+import { Colors } from "@dzcode.io/ui-mobile/dist/styles/colors";
 import React, { FC, memo } from "react";
 import { Button, Card, Paragraph, Title } from "react-native-paper";
 import { useGeneralSliceSelector } from "src/redux/reducers/general/slice";
-import { Colors } from "src/styles/colors";
 import { openLink } from "src/utils/link";
 
 import { cardStyles } from "./styles";
